requirements.txt
openpyxl==3.1.2
pre-commit==3.6.0
PyYAML==5.3.1
ruff==0.2.1
streamlit==1.29.0
streamlit-authenticator==0.1.5
